Move some unit test helpers down into BaseTestFixture
