+++ Source/CSharpEssentials.Tests/CSharpEssentials.Tests.csproj
@@ -63,6 +63,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="AnalyzerTestFixture.cs" />
+    <Compile Include="BaseTestFixture.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="UseExpressionBodiedMemberAnalyzerTests.cs" />
     <Compile Include="UseNameOfAnalyzerTests.cs" />
